feat(task-5): setting up the app for task-5
