--- package.json
@@ -1,29 +1,0 @@
-{
-  "name": "shopping-app-backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "deploy:dev": "serverless deploy --stage dev",
-    "deploy:prod": "serverless deploy --stage prod",
-    "remove:dev": "serverless remove --stage dev",
-    "remove:prod": "serverless remove --stage prod",
-    "test:fn": "serverless invoke local --function insertFakeItemsInDB"
-  },
-  "keywords": [
-    "serverless",
-    "node js",
-    "crud apis",
-    "aws rds",
-    "mysql",
-    "lambda functions",
-    "cloud formations"
-  ],
-  "author": "Ibrat Nawaz",
-  "license": "ISC",
-  "dependencies": {
-    "dotenv": "^16.0.3",
-    "mysql": "^2.18.1",
-    "uuid": "^9.0.0"
-  }
-}
