+++ product-services/.env.sample
@@ -1,0 +1,5 @@
+DB_NAME = "db_namw"
+DB_USER = "db_user"
+DB_PASSWORD = "db_password"
+AWS_RDS_ENDPOINT = "aws_db_end_point"
+DB_PORT = "db_port"
