--- resources/RoutingResources.yml
@@ -1,55 +1,0 @@
-Resources:
-  RouteTablePublic:
-    DependsOn: VPCGA
-    Type: AWS::EC2::RouteTable
-    Properties:
-      VpcId: !Ref VPC
-      Tags:
-        - Key: 'Name'
-          Value: 'RouteTablePublic'
-
-  RoutePublic:
-    Type: AWS::EC2::Route
-    Properties:
-      DestinationCidrBlock: 0.0.0.0/0
-      GatewayId: !Ref InternetGateway
-      RouteTableId: !Ref RouteTablePublic
-
-  RouteTablePrivate:
-    Type: AWS::EC2::RouteTable
-    Properties:
-      VpcId: !Ref VPC
-      Tags:
-        - Key: 'Name'
-          Value: 'RouteTablePrivate'
-
-  RoutePrivate:
-    Type: AWS::EC2::Route
-    Properties:
-      DestinationCidrBlock: 0.0.0.0/0
-      NatGatewayId: !Ref NATGateway
-      RouteTableId: !Ref RouteTablePrivate
-
-  RouteTableAssociationSubnetA:
-    Type: AWS::EC2::SubnetRouteTableAssociation
-    Properties:
-      RouteTableId: !Ref RouteTablePublic
-      SubnetId: !Ref SubnetA
-
-  RouteTableAssociationSubnetB:
-    Type: AWS::EC2::SubnetRouteTableAssociation
-    Properties:
-      RouteTableId: !Ref RouteTablePublic
-      SubnetId: !Ref SubnetB
-
-  RouteTableAssociationSubnetAPrivate:
-    Type: AWS::EC2::SubnetRouteTableAssociation
-    Properties:
-      RouteTableId: !Ref RouteTablePrivate
-      SubnetId: !Ref SubnetAPrivate
-
-  RouteTableAssociationSubnetBPrivate:
-    Type: AWS::EC2::SubnetRouteTableAssociation
-    Properties:
-      RouteTableId: !Ref RouteTablePrivate
-      SubnetId: !Ref SubnetBPrivate
